requirements.txt
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.8
neo4j==5.16.0
qdrant-client==1.12.0
ollama==0.2.1
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
networkx==3.2.1
matplotlib==3.8.2
seaborn==0.13.0
pytest==7.4.3
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.2
loguru==0.7.2
psutil==5.9.7
pyyaml==6.0.1